requirements.txt
fastapi==0.84.0
Jinja2==3.0.2
uvicorn==0.18.3
py-grpc-prometheus==0.7.0
request
kubernetes
prometheus_client
CORSMiddleware
flask